// apps/web/lib/api/models/getCalendarAccounts200ItemProvider.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * API
 * API documentation
 * OpenAPI spec version: 1.0.0
 */

export type GetCalendarAccounts200ItemProvider = typeof GetCalendarAccounts200ItemProvider[keyof typeof GetCalendarAccounts200ItemProvider];


export const GetCalendarAccounts200ItemProvider = {
  google: 'google',
} as const;